+++ lib/gossip/dissemination.js
