+++ lib/gossip/index.js
