+++ lib/gossip/join-response-merge.js
