+++ lib/gossip/join-sender.js
