+++ lib/gossip/ping-sender.js
