+++ lib/gossip/suspicion.js
